v4: Move register-async to utils/ folder. Add default for stateHandler. Fix debounce
