+++ bin/lint.sh
@@ -3,6 +3,7 @@
 npx eslint\
   ./core.js\
   ./core.test.js\
-  ./register-async.js\
+  ./utils/register-async.js\
+  ./utils/register-async.test.js\
   ./lib/\
   ./examples/
