+++ bin/transform-to-common-js.sh
@@ -8,8 +8,8 @@ file_content+="
 echo "$file_content" > __core__.js
 
 # Transforming register-async.js to CommonJS module
-file_content=`cat ./register-async.js`
+file_content=`cat ./utils/register-async.js`
 file_content="${file_content//export const/const}"
 file_content+="
 ;module.exports={registerAsync}"
-echo "$file_content" > __register-async__.js
+echo "$file_content" > utils/__register-async__.js
