+++ examples/async-search/index.html
@@ -49,5 +49,5 @@
   <body>
     <div id="root"></div>
   </body>
-  <script type="module" src="/examples/async-search/index.js"></script>
+  <script type="module" src="./index.js"></script>
 </html>
