+++ examples/async-search/ui/Root.js
@@ -1,4 +1,5 @@
 import { render } from '/core.js'
+import { debounce } from '/lib/debounce.js'
 import { dispatch } from '../store/provider.js'
 import { types } from '../types.js'
 import { SuggestionsList } from './SuggestionsList.js'
@@ -6,14 +7,16 @@ import { ErrorBanner } from './ErrorBanner.js'
 import { Spinner } from './Spinner.js'
 import { ChosenItems } from './ChosenItems.js'
 
+const onKeyup = debounce(e => {
+  dispatch({ type: types.GET_ITEMS__STARTED, query: e.target.value })
+}, 500)
+
 export const Root = () => render`
   <div id="root">
     ${ChosenItems()}
     <input
       placeholder="Search query"
-      ::keyup=${e => {
-        dispatch({ type: types.GET_ITEMS__STARTED, query: e.target.value })
-      }}
+      ::keyup=${onKeyup}
     />
     ${Spinner()}
     ${SuggestionsList()}
