+++ examples/async-todo-example/store/async-handler.js
@@ -1,4 +1,4 @@
-import { registerAsync } from '/register-async.js'
+import { registerAsync } from '/utils/register-async.js'
 
 export async function addItem(action, dispatch, state) {
   try {
