+++ examples/use-state-example/index.js
@@ -1,6 +1,6 @@
 import { createStore } from '/core.js'
 import { Root } from './Root.js'
 
-export const { mount, rerender } = createStore(() => ({}))
+export const { mount, rerender } = createStore()
 
 mount(Root)
