+++ utils/register-async.js
